refactor,profileController,reviewController

+++ controllers/user/review/reviewController.js
@@ -33,37 +33,42 @@ exports.createReview = async(req,res)=>{
     })
 }
 
-//get myReviews
-exports.getProductReview = async(req,res)=>{
-    const productId = req.params.id  
-    if(!productId){
-        return res.status(400).json({
-            message : "Please provide productId"
-        })
-    }
-    const productExist = await Product.findById(productId)
-    if(!productExist){
+//get my review
+exports.getMyReviews = async(req,res)=>{
+    const userId = req.user.id
+    const reviews = await Review.find({userId})
+    if(review.length == 0){
         return res.status(404).json({
-            message : "Product with  that id doesn't exists"
+            message : "You haven't given review to any products yet",
+            reviews : []
         })
-    }
-    const reviews = await Review.find({productId }).populate("userId")
+    }else{
         res.status(200).json({
             message : "Review fetched successfully",
-           data :  reviews
+            data : reviews
         })
+    }
 }
-
 
 //delete Review
 exports.deleteReview = async(req,res) =>{
     const reviewId = req.params.id
 
+    //check if that user created this review
     if(!reviewId){
         res.status(400).json({
             message : "Please provide reivew Id"
         })
     }
+    const userId = req.user.id
+    const review = Review.findById(reviewId)
+    const ownerOfReview = review.userId
+    if(ownerOfReview !== userId){
+        return res.status(400).json({
+            message : "You dont have permission to delete this review"
+        })
+    }
+
     await Review.findByIdAndDelete(reviewId)
     res.status(200).json({
         message : "Review deleted successfully"
@@ -71,19 +76,19 @@ exports.deleteReview = async(req,res) =>{
 }
 
 //for testing
-exports.addProductReview = async(req,res)=>{
-    const productId = req.params.id 
-    const {rating,message} = req.body 
-    const userId = req.user.id 
-    const review = {
-        userId , 
-        rating,
-        message,
-    }
-    const product = await Product.findById(productId)
-    product.reviews.push(review) 
-    await product.save() 
-    res.json({
-        message : "Review done"
-    })
-}
+// exports.addProductReview = async(req,res)=>{
+//     const productId = req.params.id 
+//     const {rating,message} = req.body 
+//     const userId = req.user.id 
+//     const review = {
+//         userId , 
+//         rating,
+//         message,
+//     }
+//     const product = await Product.findById(productId)
+//     product.reviews.push(review) 
+//     await product.save() 
+//     res.json({
+//         message : "Review done"
+//     })
+// }
